updated sign-related classes

+++ src/main/java/com/octopod/network/modules/signs/SignBukkitListener.java
@@ -1,16 +1,17 @@
 package com.octopod.network.modules.signs;
 
-import com.octopod.network.NetworkPlugin;
-import com.octopod.network.util.BukkitUtils;
+import com.octopod.network.NetworkLogger;
+import com.octopod.network.NetworkPlus;
+import com.octopod.network.cache.NetworkCommandCache;
+import com.octopod.network.commands.CommandServerConnect;
 import org.bukkit.Location;
-import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.block.Sign;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.block.Action;
 import org.bukkit.event.block.BlockBreakEvent;
-import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.block.SignChangeEvent;
+import org.bukkit.event.player.PlayerInteractEvent;
 
 /**
  * @author Octopod
@@ -18,9 +19,7 @@ import org.bukkit.event.block.SignChangeEvent;
  */
 public class SignBukkitListener implements Listener {
 
-    private String fromLoc(Location loc) {
-        return "[" + loc.getBlockX() + ", " + loc.getBlockY() + ", " + loc.getBlockZ() + ", " + loc.getWorld().getName() + "]";
-    }
+    private NetworkLogger logger = NetworkPlus.getLogger();
 
     @EventHandler
     public void onSignPlace(SignChangeEvent event) {
@@ -28,18 +27,16 @@ public class SignBukkitListener implements Listener {
         Block block = event.getBlock();
         Location loc = block.getLocation();
 
-        Sign sign = (Sign)event.getBlock().getState();
-
-        if(event.getLine(0).equals("net+sign")) {
+        if(event.getLine(0).equals("net+sign") && event.getPlayer().hasPermission("network.sign.create")) {
 
             String server = event.getLine(1);
 
-            if(NetworkPlugin.self.isServerOnline(server)) {
+            if(NetworkPlus.getInstance().isServerOnline(server)) {
                 SignPlugin.self.getDatabase().addSign(server, loc);
-                BukkitUtils.sendMessage(event.getPlayer(), "&6Net+ &7sign registered @ &e" + fromLoc(loc) + "&7!");
                 SignPlugin.self.save();
+                SignPlugin.self.updateSign(event, new SignFormat(server));
             } else {
-                BukkitUtils.sendMessage(event.getPlayer(), "&7Server &a'" + server + "'&7 doesn't exist or isn't online!");
+                logger.info("&7Server &a'" + server + "'&7 doesn't exist or isn't online!");
             }
 
         }
@@ -49,12 +46,30 @@ public class SignBukkitListener implements Listener {
     @EventHandler
     public void onBlockBreak(BlockBreakEvent event) {
 
-        Location loc =event.getBlock().getLocation();
+        Location loc = event.getBlock().getLocation();
         SignDatabase db = SignPlugin.self.getDatabase();
         if(db.getSign(loc) != null) {
-            db.removeSign(loc);
-            BukkitUtils.sendMessage(event.getPlayer(), "&6Net+ &7sign unregistered @ &e" + fromLoc(loc) + "&7!");
-            SignPlugin.self.save();
+            if(!event.getPlayer().hasPermission("network.sign.remove")) {
+                event.setCancelled(true);
+            } else {
+                db.removeSign(loc);
+                SignPlugin.self.save();
+            }
+        }
+
+    }
+
+    @EventHandler
+    public void onBlockInteract(PlayerInteractEvent event) {
+
+        if(event.getAction() == Action.RIGHT_CLICK_BLOCK) {
+
+            Location loc = event.getClickedBlock().getLocation();
+            SignDatabase db = SignPlugin.self.getDatabase();
+            if(db.getSign(loc) != null) {
+                NetworkCommandCache.getCommand(CommandServerConnect.class).onCommand(event.getPlayer(), "", new String[] {db.getSign(loc)});
+            }
+
         }
 
     }
